% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_measure.R
\name{BCD}
\alias{BCD}
\title{A measure of between components dispersion (BCD) for mixture model}
\usage{
BCD(alpha, beta, w)
}
\arguments{
\item{alpha}{loaction parameters}

\item{beta}{scale parameters}

\item{w}{weight parameters}
}
\value{
BCD
}
\description{
A measure of between components dispersion (BCD) for mixture model
}
