% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmixcauchy.R
\name{cdfmix}
\alias{cdfmix}
\title{cdfmix of the mixture of cauchy}
\usage{
cdfmix(x, mu, sigma, w)
}
\arguments{
\item{x}{Cauchy mixture data}

\item{mu}{loaction parameters}

\item{sigma}{scale parameters}

\item{w}{weight parameters}
}
\value{
cdfmix of Cauchy mixture data
}
\description{
cdfmix of the mixture of cauchy
}
