% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmixcauchy.R
\name{dmixcauchy}
\alias{dmixcauchy}
\title{density of the mixture of Cauchy}
\usage{
dmixcauchy(x, loc, scale, weight)
}
\arguments{
\item{x}{Cauchy mixture data}

\item{loc}{loaction parameters}

\item{scale}{scale parameters}

\item{weigth}{weight parameters}
}
\value{
density of Cauchy mixture data
}
\description{
density of the mixture of Cauchy
}
