% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_measure.R
\name{DOL}
\alias{DOL}
\title{Degree of overlapping (DOL)  for mixture model}
\usage{
DOL(alpha, beta, w)
}
\arguments{
\item{alpha}{loaction parameters}

\item{beta}{scale parameters}

\item{w}{weight parameters}
}
\value{
DOL
}
\description{
Degree of overlapping (DOL)  for mixture model
}
