% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_measure.R
\name{DOL_f}
\alias{DOL_f}
\title{Degree of overlapping (DOL) function for mixture model}
\usage{
DOL_f(x, alpha, beta, w)
}
\arguments{
\item{x}{Cauchy mixture data}

\item{alpha}{loaction parameters}

\item{beta}{scale parameters}

\item{w}{weight parameters}
}
\value{
DOL function
}
\description{
Degree of overlapping (DOL) function for mixture model
}
