% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmixcauchy.R
\name{pdfmix}
\alias{pdfmix}
\title{pdfmix of the mixture of Cauchy}
\usage{
pdfmix(x, mu, sigma, w)
}
\arguments{
\item{x}{Cauchy mixture data}

\item{mu}{loaction parameters}

\item{sigma}{scale parameters}

\item{w}{weight parameters}
}
\value{
pdfmix of Cauchy mixture data
}
\description{
pdfmix of the mixture of Cauchy
}
