% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cauchy.R
\name{plot_cauchy}
\alias{plot_cauchy}
\title{Histogram of Cauchy mixture with true paramter}
\usage{
plot_cauchy(x, mu, sigma, lambda, a, b)
}
\arguments{
\item{x}{mixture data}

\item{mu}{lcoation parameters}

\item{sigma}{scale parameters}

\item{lambda}{weight parameters}

\item{a}{plot start from}

\item{b}{plot end with}
}
\value{
histogram
}
\description{
Histogram of Cauchy mixture with true paramter
}
