% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmixcauchy.R
\name{pmixcauchy}
\alias{pmixcauchy}
\title{distribution of the mixture of cauchy}
\usage{
pmixcauchy(x, loc, scale, weight)
}
\arguments{
\item{x}{Cauchy mixture data}

\item{loc}{loaction parameters}

\item{scale}{scale parameters}

\item{weight}{weight parameters}
}
\value{
distribution of Cauchy mixture data
}
\description{
distribution of the mixture of cauchy
}
