% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixcauchy.R
\name{rmixcauchy}
\alias{rmixcauchy}
\title{Cauchy mixture data generation}
\usage{
rmixcauchy(n, loc, scale, weight, seed)
}
\arguments{
\item{n}{number of sample}

\item{loc}{loaction parameters}

\item{scale}{scale parameters}

\item{seed}{random seed}

\item{weigth}{weight parameters}
}
\value{
Cauchy mixture data
}
\description{
Cauchy mixture data generation
}
